Change Base ORM class
The original Base setup wasn't flexible enough to allow single
inheritance on the Permission table. I changed the Base class to an
empty class and moved the id column and table_name function
to a `Default` class, which can be used as a Mixin for most tables.
This allowed setting NodePermission table to have no table_name or
id column, which is necessary setup for single inheritance.