Add user group test for fail case
Fails if the resource whose permissions are being granted to the
user group is not descended from the user group's parent node in the
node tree.

Add user group endpoint to pass this test.

Add function to app.crud.base to fetch node trees starting with a
root node

Add functions to AccessControl class to get node descendants and check
if a target resource is in the descendant tree (abstract)

Implemented abstract is_descended_from function in CRUDNode, still need
to do it for CRUDUserGroup