Update model_encoder
Previous version couldn't handle joined table inheritance in SQLAlchemy
models. The current solution is to force a refresh on the object prior
to encoding if a session is also passed to model_encoder. This is really
the solution I would have preferred all along, I just didn't realize
before that calling db.refresh() on an object would force the object
fields to load.