Add fields to QueryInterface model
Added fields to hold the last fetched result (last_result), to indicate
the timestamp of the last run (laast_run), to indicate a time in seconds
before the query result needs to be refreshed (refresh_interval), and to
hold the total rows in the full query, in case the user is paginating
through the results.

Also changed the JSON columns to JSONB columns for performance sake.